stabilize the use of the self_intersection_plugin by moving it to the selection_plugin
