+++ Polyhedron/demo/Polyhedron/CMakeLists.txt
@@ -362,6 +362,11 @@ add_executable  ( CGAL_Classification Classification.cpp )
 add_dependencies(CGAL_Classification Classification)
 target_link_libraries( CGAL_Classification PRIVATE polyhedron_demo )
 add_to_cached_list( CGAL_EXECUTABLE_TARGETS CGAL_Classification )
+
+add_executable  ( CGAL_PMP PMP.cpp )
+add_dependencies(CGAL_PMP PMP)
+target_link_libraries( CGAL_PMP PRIVATE polyhedron_demo )
+add_to_cached_list( CGAL_EXECUTABLE_TARGETS CGAL_PMP )
 #
 # Exporting
 #
