+++ Polyhedron/demo/Polyhedron/Plugins/PMP/CMakeLists.txt
@@ -12,14 +12,14 @@ if(EIGEN3_FOUND)
   if("${EIGEN3_VERSION}" VERSION_GREATER "3.1.90")
 
     qt5_wrap_ui( hole_fillingUI_FILES Hole_filling_widget.ui)
-    polyhedron_demo_plugin(hole_filling_plugin Hole_filling_plugin ${hole_fillingUI_FILES})
+    polyhedron_demo_plugin(hole_filling_plugin Hole_filling_plugin ${hole_fillingUI_FILES} KEYWORDS PMP)
     target_link_libraries(hole_filling_plugin PUBLIC scene_surface_mesh_item scene_polylines_item scene_selection_item)
 
     qt5_wrap_ui( fairingUI_FILES Fairing_widget.ui)
-    polyhedron_demo_plugin(fairing_plugin Fairing_plugin ${fairingUI_FILES})
+    polyhedron_demo_plugin(fairing_plugin Fairing_plugin ${fairingUI_FILES} KEYWORDS PMP)
     target_link_libraries(fairing_plugin PUBLIC scene_selection_item)
 
-    polyhedron_demo_plugin(hole_filling_polyline_plugin Hole_filling_polyline_plugin)
+    polyhedron_demo_plugin(hole_filling_polyline_plugin Hole_filling_polyline_plugin )
     target_link_libraries(hole_filling_polyline_plugin PUBLIC scene_surface_mesh_item scene_polylines_item)
     
     qt5_wrap_ui( Mean_curvature_flow_skeleton_pluginUI_FILES  Mean_curvature_flow_skeleton_plugin.ui)
@@ -39,14 +39,14 @@ else()
 endif()
 
 qt5_wrap_ui( soupUI_FILES Repair_soup.ui )
-polyhedron_demo_plugin(orient_soup_plugin Orient_soup_plugin ${soupUI_FILES} KEYWORDS Classification )
+polyhedron_demo_plugin(orient_soup_plugin Orient_soup_plugin ${soupUI_FILES} KEYWORDS Classification PMP)
 target_link_libraries(orient_soup_plugin PUBLIC scene_polygon_soup_item scene_surface_mesh_item scene_polylines_item scene_points_with_normal_item)
 
 
-polyhedron_demo_plugin(inside_out_plugin Inside_out_plugin)
+polyhedron_demo_plugin(inside_out_plugin Inside_out_plugin KEYWORDS PMP)
 target_link_libraries(inside_out_plugin PUBLIC scene_surface_mesh_item scene_polygon_soup_item)
 
-polyhedron_demo_plugin(join_and_split_plugin Join_and_split_polyhedra_plugin)
+polyhedron_demo_plugin(join_and_split_plugin Join_and_split_polyhedra_plugin KEYWORDS PMP)
 target_link_libraries(join_and_split_plugin PUBLIC scene_surface_mesh_item scene_selection_item)
 
 
@@ -60,58 +60,59 @@ qt5_wrap_ui( polyhedron_slicerUI_FILES Polyhedron_slicer_widget.ui)
 polyhedron_demo_plugin(polyhedron_slicer_plugin Polyhedron_slicer_plugin ${polyhedron_slicerUI_FILES})
 target_link_libraries(polyhedron_slicer_plugin PUBLIC scene_surface_mesh_item scene_basic_objects scene_polylines_item)
 
-polyhedron_demo_plugin(polyhedron_stitching_plugin Polyhedron_stitching_plugin)
+polyhedron_demo_plugin(polyhedron_stitching_plugin Polyhedron_stitching_plugin KEYWORDS PMP)
 target_link_libraries(polyhedron_stitching_plugin PUBLIC scene_surface_mesh_item scene_polylines_item)
 
 qt5_wrap_ui( selectionUI_FILES Selection_widget.ui)
-polyhedron_demo_plugin(selection_plugin Selection_plugin ${selectionUI_FILES} KEYWORDS PolygonMesh IO Classification Mesh_3)
+polyhedron_demo_plugin(selection_plugin Selection_plugin ${selectionUI_FILES} KEYWORDS PMP IO Classification Mesh_3)
 target_link_libraries(selection_plugin PUBLIC scene_selection_item scene_points_with_normal_item scene_polylines_item)
 
-polyhedron_demo_plugin(self_intersection_plugin Self_intersection_plugin)
-target_link_libraries(self_intersection_plugin PUBLIC scene_selection_item scene_surface_mesh_item)
+#to keep it simple to compile
+add_custom_target(self_intersection_plugin )
+add_dependencies(self_intersection_plugin selection_plugin)
 
-polyhedron_demo_plugin(triangulate_facets_plugin Triangulate_facets_plugin)
+polyhedron_demo_plugin(triangulate_facets_plugin Triangulate_facets_plugin KEYWORDS PMP)
 target_link_libraries(triangulate_facets_plugin PUBLIC scene_surface_mesh_item)
 
-polyhedron_demo_plugin(corefinement_plugin Corefinement_plugin)
+polyhedron_demo_plugin(corefinement_plugin Corefinement_plugin KEYWORDS PMP)
 target_link_libraries(corefinement_plugin PUBLIC scene_surface_mesh_item)
 
-polyhedron_demo_plugin(surface_intersection_plugin Surface_intersection_plugin)
+polyhedron_demo_plugin(surface_intersection_plugin Surface_intersection_plugin KEYWORDS PMP)
 target_link_libraries(surface_intersection_plugin PUBLIC scene_surface_mesh_item scene_polylines_item scene_points_with_normal_item)
 
-polyhedron_demo_plugin(repair_polyhedron_plugin Repair_polyhedron_plugin)
+polyhedron_demo_plugin(repair_polyhedron_plugin Repair_polyhedron_plugin KEYWORDS PMP)
 target_link_libraries(repair_polyhedron_plugin PUBLIC scene_surface_mesh_item)
 
 qt5_wrap_ui( isotropicRemeshingUI_FILES Isotropic_remeshing_dialog.ui)
-polyhedron_demo_plugin(isotropic_remeshing_plugin Isotropic_remeshing_plugin  ${isotropicRemeshingUI_FILES})
+polyhedron_demo_plugin(isotropic_remeshing_plugin Isotropic_remeshing_plugin  ${isotropicRemeshingUI_FILES} KEYWORDS PMP)
 target_link_libraries(isotropic_remeshing_plugin PUBLIC scene_surface_mesh_item scene_selection_item)
 
 if(TBB_FOUND)
   CGAL_target_use_TBB(isotropic_remeshing_plugin)
 endif()
 
-polyhedron_demo_plugin(distance_plugin Distance_plugin)
+polyhedron_demo_plugin(distance_plugin Distance_plugin KEYWORDS PMP)
 target_link_libraries(distance_plugin PUBLIC scene_surface_mesh_item  scene_color_ramp)
 if(TBB_FOUND)
   CGAL_target_use_TBB(distance_plugin)
 endif()
 
 
-polyhedron_demo_plugin(detect_sharp_edges_plugin Detect_sharp_edges_plugin KEYWORDS IO Mesh_3)
+polyhedron_demo_plugin(detect_sharp_edges_plugin Detect_sharp_edges_plugin KEYWORDS IO Mesh_3 PMP)
 target_link_libraries(detect_sharp_edges_plugin PUBLIC scene_surface_mesh_item)
 
 
 
 qt5_wrap_ui( randomPerturbationUI_FILES Random_perturbation_dialog.ui)
-polyhedron_demo_plugin(random_perturbation_plugin Random_perturbation_plugin ${randomPerturbationUI_FILES})
+polyhedron_demo_plugin(random_perturbation_plugin Random_perturbation_plugin ${randomPerturbationUI_FILES} KEYWORDS PMP)
 target_link_libraries(random_perturbation_plugin PUBLIC scene_surface_mesh_item scene_selection_item)
 
-polyhedron_demo_plugin(degenerated_faces_plugin Degenerated_faces_plugin)
+polyhedron_demo_plugin(degenerated_faces_plugin Degenerated_faces_plugin KEYWORDS PMP)
 target_link_libraries(degenerated_faces_plugin PUBLIC scene_surface_mesh_item scene_selection_item)
 
 qt5_wrap_ui( engravUI_FILES Engrave_dock_widget.ui )
 polyhedron_demo_plugin(engrave_text_plugin Engrave_text_plugin ${engravUI_FILES})
 target_link_libraries(engrave_text_plugin PUBLIC scene_surface_mesh_item scene_selection_item scene_polylines_item)
 
-polyhedron_demo_plugin(extrude_plugin Extrude_plugin)
+polyhedron_demo_plugin(extrude_plugin Extrude_plugin KEYWORDS PMP)
 target_link_libraries(extrude_plugin PUBLIC scene_surface_mesh_item scene_selection_item)
